refactor(memory): split semantic-memory-update-service into sub-modules
Decompose the 1168-line SemanticMemoryUpdateService god node into focused
composition modules while preserving the public API and all 45 vitest cases.

Closes #598
